Initialized the Intake Motors, Programmed Intake, and Swapped Precise and Normal Mode Buttons
Initialized the Bottom and Top Half Intake Motors and Programmed the Intake for the Bottom and Top Middle Goals and a Stop Intake Function. Also Remapped the Precise Mode Button to Share Button Y With the Normal Mode, to Create a Mode Swap from One Button (Press One to Go to Precise Mode, Press Again to Go to Normal Mode).